finishing rest of the part 4 exercises

+++ osa4/controllers/blogs.js
@@ -2,14 +2,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog.js')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+const helper = require('../tests/test_helper.js')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -19,11 +12,10 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'not authorized' })
   }
-  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: body.title,
@@ -40,8 +32,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  const user = request.user
+  if (!user) {
+    return response.status(404).json({ error: 'user not found' })
+  }
+
+  if (blog.user.toString() === user.id.toString()){
+    await Blog.findByIdAndDelete(request.params.id)
+    return response.status(204).end()
+  }
+  else{
+    return response.status(401).json({ error: 'not authorized to delete this blog' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) =>{
